+++ src/Endpoints/Email.php
@@ -60,6 +60,7 @@ class Email extends AbstractEndpoint
                         'track_opens' => $params->trackOpens(),
                         'track_content' => $params->trackContent(),
                     ],
+                    'headers' => $params->getHeaders(),
                 ],
                 fn ($v) => is_array($v) ? array_filter($v, fn ($v) => $v !== null) : $v !== null
             )