+++ CHANGELOG.md
@@ -15,6 +15,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.19.0] - 2023 12 19
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
+- Remove symfony dependency
+
+## [0.18.0] - 2023 12 13
+
+### Added
+
+- Add an endpoint to fetch a single email activity
+
+### Changed
+
+### Fixed
+
+### Removed
+
+## [0.17.0] - 2023 12 12
+
+### Added
+
+- Allow user to send custom headers
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [0.16.0] - 2023 11 22
 
 ### Added